Implement md5 digest
By the original implementation of eBay, they support only sha256 and sha512.
But in our case, we can benefit from md5 that is already being computed for all files out of box by AWS S3 when the file is uploaded.
We don't care much about the stability of the cipher algorithm for digest because there is not too much files being signed here.
MD5 is jsut fine and fast and really lightweight.